[FEAT/#121] 등록 다이얼로그 백핸들러 추가
