+++ app/src/main/java/com/napzak/market/presentation/home/HomeScreen.kt
@@ -1,7 +1,5 @@
 package com.napzak.market.presentation.home
 
-import android.app.Activity
-import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.LocalOverscrollConfiguration
 import androidx.compose.foundation.background
@@ -10,28 +8,16 @@ import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.SnackbarDuration
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableLongStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.napzak.market.R.string.home_snack_bar_finish
 import com.napzak.market.core.common.state.UiState
-import com.napzak.market.core.designsystem.component.snackbar.CommonSnackBar
 import com.napzak.market.core.designsystem.component.topbar.NapzakLogoTopBar
 import com.napzak.market.core.designsystem.theme.NapzakMarketTheme
 import com.napzak.market.core.type.TradeType
@@ -43,7 +29,6 @@ import com.napzak.market.presentation.home.component.HomePopularItemGroup
 import com.napzak.market.presentation.home.component.HomeRecommendationItemGroup
 import com.napzak.market.presentation.home.state.HomeUiState
 import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.launch
 
 @Composable
 fun HomeRoute(
@@ -76,29 +61,6 @@ private fun HomeScreen(
     onItemClick: (Long) -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    val context = LocalContext.current
-    var backPressedState by remember { mutableStateOf(true) }
-    var backPressedTime by remember { mutableLongStateOf(0L) }
-
-    val snackBarHostState = remember { SnackbarHostState() }
-    val coroutineScope = rememberCoroutineScope()
-
-    BackHandler(enabled = backPressedState) {
-        if (System.currentTimeMillis() - backPressedTime <= 2000) {
-            (context as Activity).finish()
-        } else {
-            backPressedState = true
-            coroutineScope.launch {
-                snackBarHostState.showSnackbar(
-                    message = context.getString(home_snack_bar_finish),
-                    duration = SnackbarDuration.Short
-                )
-            }
-        }
-        backPressedTime = System.currentTimeMillis()
-    }
-
-
     CompositionLocalProvider(
         value = LocalOverscrollConfiguration provides null,
     ) {
@@ -152,20 +114,6 @@ private fun HomeScreen(
                         }
                     )
                 }
-            }
-
-            SnackbarHost(
-                hostState = snackBarHostState,
-                modifier = Modifier
-                    .align(Alignment.BottomCenter)
-                    .padding(bottom = 6.dp),
-            ) { snackBarData ->
-                CommonSnackBar(
-                    message = snackBarData.visuals.message,
-                    backgroundColor = NapzakMarketTheme.colors.black70,
-                    textColor = NapzakMarketTheme.colors.white,
-                    textStyle = NapzakMarketTheme.typography.bodyMedium14,
-                )
             }
         }
     }
